% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plans.R
\name{add_test_plan}
\alias{add_test_plan}
\title{Add a Future Plan to Test Against}
\usage{
add_test_plan(expr, substitute = TRUE)
}
\arguments{
\item{expr}{...}

\item{substitute}{...}
}
\value{
(invisibly) returns current list of test plans.
}
\description{
Add a Future Plan to Test Against
}
