% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{along_cores}
\alias{along_cores}
\title{Evaluate Tests Emulating One, Two, ... Cores}
\usage{
along_cores(expr, envir = parent.frame())
}
\arguments{
\item{expr}{An \R expression.}

\item{envir}{The environment from where tests are run.}
}
\value{
(invisible) A named list where each element holds the
value of the expression evaluated with a given number of cores.
}
\description{
Evaluate Tests Emulating One, Two, ... Cores
}
