% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{along_strategies}
\alias{along_strategies}
\title{Evaluate Tests Across A Set of Future Strategies}
\usage{
along_strategies(expr, envir = parent.frame())
}
\arguments{
\item{expr}{An \R expression.}

\item{envir}{The environment from where tests are run.}
}
\value{
(invisible) A named list where each element holds the
value of the expression evaluated with a given strategy.
}
\description{
Evaluate Tests Across A Set of Future Strategies
}
