% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/along_test_plans.R
\name{along_test_plans}
\alias{along_test_plans}
\title{Evaluate an Expression Across A Set of Future Plans}
\usage{
along_test_plans(
  expr,
  substitute = TRUE,
  envir = parent.frame(),
  local = TRUE,
  plans = test_plans()
)
}
\arguments{
\item{expr}{An \R expression.}

\item{substitute}{...}

\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{plans}{A list of future plans.}
}
\value{
A list of results, one for each future plan tested against.
}
\description{
Evaluate an Expression Across A Set of Future Plans
}
