% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{assert_future_explicit}
\alias{assert_future_explicit}
\title{Assert that future and regular evaluation give identical values}
\usage{
assert_future_explicit(expr, prepend = NULL, envir = parent.frame(), ...,
  substitute = TRUE)
}
\arguments{
\item{expr}{An \R \link[base]{expression} to be evaluated.}

\item{prepend}{(optional) An expression prepended to the future expression.}

\item{envir}{The \link{environment} from where global objects should be
identified.  Depending on the future strategy (the \code{evaluator}), it may
also be the environment in which the expression is evaluated.}

\item{\dots}{(optional) Additional arguments pass to
\code{\link[future:future]{future()}.}}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}
}
\value{
A named \link{list}.
}
\description{
Assert that future and regular evaluation give identical values
}
