% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{begin}
\alias{begin}
\alias{end}
\alias{context}
\title{Setup and Teardown a Package Test}
\usage{
begin(title, package = "future", attach = TRUE, envir = parent.frame())

end()

context(msg)
}
\arguments{
\item{title}{A character string for the test title.}

\item{package}{Name of package tests.}

\item{attach}{If TRUE, the package \code{pkg} is attached, otherwise
it only loaded.}

\item{envir}{The environment from where tests are run.}

\item{msg}{A character string.}
}
\value{
Nothing.
}
\description{
Setup and Teardown a Package Test
}
