% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Run All Tests Across Future Plans}
\usage{
check(args = commandArgs())
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\value{
Nothing.
}
\description{
Run All Tests Across Future Plans
}
\section{Command-line interface (CLI)}{

Some examples on how to call this function from the command line:
\preformatted{
Rscript -e future.tests::check --args --test-plan=sequential
Rscript -e future.tests::check --args --test-plan=multicore,workers=2
Rscript -e future.tests::check --args --test-plan=sequential --test-plan=multicore,workers=2
}
}

