% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plan.R
\name{check_plan}
\alias{check_plan}
\title{Run All Tests}
\usage{
check_plan(
  tests = test_db(),
  defaults = list(),
  timeout = getOption("future.tests.timeout", 30),
  envir = parent.frame()
)
}
\arguments{
\item{tests}{A list of tests to subset.}

\item{defaults}{(optional) Named list with default argument values.}

\item{timeout}{Maximum time allowed for evaluation before a timeout error is produced.}

\item{envir}{The environment where tests are run.}
}
\value{
Nothing.
}
\description{
Run All Tests
}
