% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test-class.R
\name{evaluate_expr}
\alias{evaluate_expr}
\title{Run a Test}
\usage{
evaluate_expr(expr, envir = parent.frame(), local = TRUE)
}
\arguments{
\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{test}{A Test.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Run a Test
}
