% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_options}
\alias{get_options}
\title{Gets R Options by Name Matching}
\usage{
get_options(pattern = NULL, fixed = FALSE)
}
\arguments{
\item{pattern}{A regular expression or a fixed string.}

\item{fixed}{If TRUE, the \code{pattern} is matched as a fixed string,
otherwise as a regular expression.}
}
\value{
A named list of zero or more options.
}
\description{
Gets R Options by Name Matching
}
