% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_solaris}
\alias{is_solaris}
\title{Checks Whether Tests are Run on Solaris or not}
\usage{
is_solaris()
}
\value{
TRUE if running on Solaris, otherwise FALSE
}
\description{
Checks Whether Tests are Run on Solaris or not
}
