% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_db.R
\name{load_tests}
\alias{load_tests}
\title{Loads Future Tests}
\usage{
load_tests(
  path = ".",
  recursive = TRUE,
  pattern = "[.]R$",
  root = getOption("future.tests.root", Sys.getenv("R_FUTURE_TESTS_ROOT",
    system.file("test-db", package = "future.tests", mustWork = TRUE)))
)
}
\arguments{
\item{path}{A character string specifying a test script folder or file.}

\item{recursive}{If TRUE, test-definition scripts are search recursively.}

\item{pattern}{Regular expression matching filenames to include.}

\item{root}{(internal) An alternative file directory from where
\pkg{future.tests} tests are sourced.}
}
\value{
Number of tests added.
}
\description{
Loads Future Tests
}
