% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test-class.R
\name{make_test}
\alias{make_test}
\title{Make a Test}
\usage{
make_test(.title = NA_character_, ..., .expr, .substitute = TRUE,
  .register = TRUE)
}
\arguments{
\item{.title}{(character) The title of the test.}

\item{.expr, .substitute}{The expression to be tested and
whether it is passes as an expression already or not.}

\item{.register}{If TRUE, the test is registered in the test database, otherwise not.}

\item{\ldots}{(optional) Named arguments.}
}
\value{
(invisibly) A Test.
}
\description{
Make a Test
}
