% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test-class.R
\name{run_test}
\alias{run_test}
\title{Run a Test}
\usage{
run_test(test, envir = parent.frame())
}
\arguments{
\item{test}{A Test.}

\item{envir}{The environment where tests are run.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Run a Test
}
