% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_test.R
\name{skip_test}
\alias{skip_test}
\title{Skip The Current Test}
\usage{
skip_test(message = "Test skipped", domain = NULL)
}
\arguments{
\item{domain}{see \code{\link[base]{gettext}}.  If \code{NA}, messages will
    not be translated, see also the note in \code{\link[base]{stop}}.}
}
\value{
(invisible) A \link[base:conditions]{base::condition} of class \code{TestSkipped}.
}
\description{
Signals a \code{TestSkipped} condition.
}
