% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{strategies}
\alias{strategies}
\title{Get and Set Future Strategies to Test Over}
\usage{
strategies(cores = availableCores(), excl = c("multiprocess",
  "cluster"), ...)
}
\arguments{
\item{cores}{(integer) The number of cores.}

\item{excl}{A character vector of strategies to \emph{not} test.}

\item{\ldots}{(optional) Arguments passed to
\code{future:::supportedStrategies()}.}
}
\value{
A character vector of zero or more future strategies.
}
\description{
Get and Set Future Strategies to Test Over
}
