% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_db.R
\name{subset_tests_by_args}
\alias{subset_tests_by_args}
\title{Identify Subset of Tests that Support Specified Argument Settings}
\usage{
subset_tests_by_args(tests = test_db(), tags = NULL, args = NULL)
}
\arguments{
\item{tests}{A list of tests to subset.}

\item{tags}{(optional) A character vector of tags that tests must have.}

\item{args}{Named arguments with sets of values to test against.}
}
\value{
A list of tests that support specified arguments.
}
\description{
Identify Subset of Tests that Support Specified Argument Settings
}
