% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{test}
\alias{test}
\title{Run a Future Test}
\usage{
test(path, ..., root = getOption("future.tests.root",
  Sys.getenv("R_FUTURE_TESTS_ROOT", system.file("tests", package =
  "future.tests", mustWork = TRUE))))
}
\arguments{
\item{path}{A character string specifying a test script.}

\item{root}{(internal) An alternative file directory from where
\pkg{future.tests} tests are sourced.}

\item{\ldots}{Additional arguments passed to \link[base:source]{base::source}.}
}
\value{
Nothing
}
\description{
Run a Future Test
}
