% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_db.R
\name{test_db}
\alias{test_db}
\title{Manage the Test Database}
\usage{
test_db(action = c("list", "append", "reset"), test = NULL)
}
\arguments{
\item{action}{(character) What should be done.}

\item{test}{A Test.}
}
\value{
(invisibly) the internal list of Test:s.
}
\description{
Manage the Test Database
}
\keyword{internal}
