% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plan.R
\name{test_plans}
\alias{test_plans}
\title{Manage the Set of Future Plans to Test Against}
\usage{
test_plans(action = c("list", "add"), expr = NULL, substitute = TRUE)
}
\arguments{
\item{action}{...}

\item{expr}{...}

\item{substitute}{...}
}
\value{
(invisibly) returns current list of test plans.
}
\description{
Manage the Set of Future Plans to Test Against
}
\keyword{internal}
