% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Run All Tests}
\usage{
check(tests = test_db(), defaults = list(),
  timeout = getOption("future.tests.timeout", 30))
}
\arguments{
\item{tests}{A list of tests to subset.}

\item{defaults}{(optional) Named list with default argument values.}

\item{timeout}{Maximum time allowed for evaluation before a timeout error is produced.}
}
\value{
Nothing.
}
\description{
Run All Tests
}
