% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_expr.R
\name{evaluate_expr}
\alias{evaluate_expr}
\title{Evaluate an R Expression}
\usage{
evaluate_expr(expr, envir = parent.frame(), local = TRUE,
  stdout = TRUE)
}
\arguments{
\item{expr}{An expression}

\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{stdout}{If TRUE, standard output is captured, otherwise not.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Evaluate an R Expression
}
