% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_expr.R
\name{evaluate_expr}
\alias{evaluate_expr}
\title{Evaluate an R Expression}
\usage{
evaluate_expr(
  expr,
  envir = parent.frame(),
  local = TRUE,
  output = c("stdout+stderr", "stdout", "none"),
  timeout = +Inf
)
}
\arguments{
\item{expr}{An expression}

\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{output}{Specifies whether standard output, standard error, or both should be captured or not.}

\item{timeout}{Maximum time allowed for evaluation before a timeout error is produced.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Evaluate an R Expression
}
