% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test.R
\name{make_test}
\alias{make_test}
\title{Make a Test}
\usage{
make_test(
  expr,
  title = NA_character_,
  args = list(),
  tags = NULL,
  substitute = TRUE,
  reset_workers = FALSE,
  register = TRUE
)
}
\arguments{
\item{expr, substitute}{The expression to be tested and
whether it is passes as an expression already or not.}

\item{title}{(character) The title of the test.}

\item{args}{(optional) Named arguments.}

\item{tags}{(optional) Character vector of tags.}

\item{reset_workers}{(optional) Specifies whether background workers should
futures.}

\item{register}{If TRUE, the test is registered in the test database,
otherwise not.}
}
\value{
(invisibly) A Test.
}
\description{
Make a Test
}
