% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestResult-class.R
\name{run_test}
\alias{run_test}
\title{Run a Test}
\usage{
run_test(test, envir = parent.frame(), local = TRUE, stdout = TRUE)
}
\arguments{
\item{test}{A Test.}

\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{stdout}{If TRUE, standard output is captured, otherwise not.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Run a Test
}
