% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{Run a Test}
\usage{
run_test(
  test,
  envir = parent.frame(),
  local = TRUE,
  args = list(),
  defaults = list(),
  output = "stdout+stderr",
  timeout = getOption("future.tests.timeout", 30)
)
}
\arguments{
\item{test}{A Test.}

\item{envir}{The environment where tests are run.}

\item{local}{Should tests be evaluated in a local environment or not.}

\item{args}{Arguments used in this test.}

\item{defaults}{(optional) Named list with default argument values.}

\item{output}{If TRUE, standard output is captured, otherwise not.}

\item{timeout}{Maximum time allowed for evaluation before a timeout error is produced.}
}
\value{
Value of test expression and benchmark information.
}
\description{
Run a Test
}
