% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{test}
\alias{test}
\title{Runs a Future Test}
\usage{
test(path, ...)
}
\arguments{
\item{path}{A character string specifying a test script.}

\item{\ldots}{Additional arguments passed to \link[base:source]{base::source}.}
}
\value{
Nothing
}
\description{
Runs a Future Test
}
